% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av-transform.R
\name{av_labs}
\alias{av_labs}
\title{Labels for an added-variable plot}
\usage{
av_labs(data)
}
\arguments{
\item{data}{A data frame returned by \code{\link[=av_transform]{av_transform()}}.}
}
\value{
A ggplot2 labels object.
}
\description{
\code{av_labs()} returns ggplot2 axis labels that describe the residual models used
by \code{\link[=av_transform]{av_transform()}}.
}
\examples{
av_data <- av_transform(mtcars, y = mpg, x = wt, adjust = c(hp, factor(cyl)))

ggplot2::ggplot(av_data, ggplot2::aes(.adjusted_wt, .adjusted_mpg)) +
  ggplot2::geom_point() +
  av_labs(av_data)
}
