% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{honestlm-package}
\alias{honestlm}
\alias{honestlm-package}
\title{honestlm: Guardrails for Teaching Linear Models}
\description{
Provides small, opinionated wrappers around familiar linear model workflows. The package keeps base R and ggplot2 interfaces recognizable while adding guardrails around common interpretation traps, including coefficient p-values for multi-level categorical predictors, sequential Type I sums of squares, and plots whose summaries do not match the fitted linear model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yanivjb.github.io/honestlm/}
  \item \url{https://github.com/yanivjb/honestlm}
  \item Report bugs at \url{https://github.com/yanivjb/honestlm/issues}
}

}
\author{
\strong{Maintainer}: Yaniv Brandvain \email{ybrandva@umn.edu}

Authors:
\itemize{
  \item Yaniv Brandvain \email{ybrandva@umn.edu}
}

}
\keyword{internal}
