% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.honest_lm}
\alias{anova.honest_lm}
\title{Analysis of variance for a guarded linear model}
\usage{
\method{anova}{honest_lm}(object, ...)
}
\arguments{
\item{object}{An \code{honest_lm} object.}

\item{...}{Passed to \code{\link[stats:anova]{stats::anova()}}.}
}
\value{
An ANOVA table with an extra class for printing.
}
\description{
This method returns the same sequential ANOVA table as \code{\link[stats:anova.lm]{stats::anova.lm()}},
but warns that these are Type I sums of squares and therefore depend on term
order.
}
